code review fixes

+++ java/prepare.sh
@@ -6,7 +6,7 @@ echo "Download openapi file"
 curl -o openapi.json https://piiano.com/docs/assets/openapi.json
 
 echo "Delete the old generated java SDK"
-rm -r vault_java_sdk
+[ -d "vault_java_sdk" ] && rm -r vault_java_sdk
 
 echo "Run openapi tools to create the java SDK"
 docker run --rm -u $(id -u):$(id -g) -v "${PWD}:/local" openapitools/openapi-generator-cli:v6.2.1 generate \
